refactor :: remove type
- 불필요한 tpye 제거
- Props type은 해당 파일에서 사용하도록 변경
