+++ src/types/auth/index.ts
@@ -1,2 +1,3 @@
 export * from "./signup";
 export * from "./login";
+export * from "./user";
