+++ src/types/index.ts
@@ -1,5 +1,4 @@
 export * from "./auth";
-export * from "./main";
 export * from "./ai";
 export * from "./affiliation";
 export * from "./member";
