--- src/types/main/chatbot.ts
@@ -1,4 +1,0 @@
-export interface ChatCardProps {
-  text: string;
-  who: "user" | "ai";
-}
