--- src/types/main/index.ts
@@ -1,2 +1,0 @@
-export * from "./ranking";
-export * from "./user";
