--- src/types/main/ranking.ts
@@ -1,8 +1,0 @@
-import { ImageSourcePropType } from "react-native";
-
-export interface RankUserProps {
-  rank: number;
-  profileImage: string;
-  profileName: string;
-  bookPoint: number;
-}
